Add some setup details to README
Added some installation details to the README.

Picking the target for the Raspberry Pi Zero 2 series --- which we're
using as the default target for this application --- isn't documented in
the official docs yet, so we note it here.

Since `fwup` isn't included in the default Ubuntu 22.04 repos, we've
added a link to the installation docs for future reference.